10-langchain-langsmith-fastapi/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0

langchain>=0.2.0
langchain-openai>=0.1.0
langsmith>=0.1.0

python-dotenv>=1.0.0
